+++ packages/freezed/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased patch
+
+- Stop using `JsonKey(ignore: true)` in favour of `JsonKey(includeFromJson: false, includeToJson: false)`
+  (thanks to @lrsvmb)
+- Require json_annotation ^6.8.0
+
 ## 2.5.4 - 2024-07-02
 
 - Require Dart >=3.0.0